requirements.txt
yfinance
matplotlib
requests
py_vollib
numpy
scipy
flask
flask-cors
python-dotenv
GitPython